requirements_streamlit.txt
huggingface_hub>=0.23
pyyaml
requests
streamlit
